Write speed values to Google Sheet
Added functionality to write the speed test results to a Google Sheet.
The required client_id.json files are not included here and should be
generated using the Google Developer Console.

Currently in progress. Issues: a new sheet is created each time the
script is run as opposed to adding a new row to the worksheet.