Append new results to existing sheet
Checks if a spreadsheet was previously created for the speedtest
results. If not, creates a new sheet in the linked Google Drive and
stores the spreadsheet id in a file locally.

If a spreadsheet was created previously, use that as the spreadsheet to
work on.

New results are now appended to an existing spreadsheet.

Code was also refactored quite a bit to improve functionality.